requirements.txt
Flask==2.1.0
Werkzeug==2.0.3
pymongo==4.3.3
flask-compress==1.13
nltk==3.8.1
scikit-learn==1.3.0
numpy==1.24.3
python-dotenv==1.0.0
certifi==2023.7.22
groq==0.4.0
pytz==2023.3
uuid==1.30
bson==0.5.10

# Face recognition and computer vision
face-recognition==1.3.0
opencv-python==4.8.0.74
Pillow==9.5.0

# Google Fit API integration
google-auth-oauthlib==1.0.0
google-auth==2.22.0
google-api-python-client==2.95.0

# YOLOv8 dependencies (for computer vision)
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.145

# Web related utilities
requests==2.31.0
